add ONLY_MAP option to encode.sh
I wanted to audit the configs to ensure they didn't create any overlapping output files.
The logic for determining the output path is fairly complex, though, and lives in encode.sh.
It's much simpler at this point to just add a variable to have the script spit out the paths
and exit when it's done. Then a bit of shell scripting can easily grab all the output paths
and look for overlaps.

+++ allinone/encode.sh
@@ -238,6 +238,12 @@ function encode_one {
 
     output_abs_path="$base_output_dir/$output_rel_path"
     output_tmp_path="$base_output_dir/$(dirname "$output_rel_path")/$(basename "$output_rel_path").part"
+
+    if [ -n "$ONLY_MAP" ]; then
+        echo "IN: $input_abs_path OUT: $output_abs_path"
+        return 0
+    fi
+
     trap 'rm -f "$output_tmp_path"' EXIT
 
     DONEFILE="$DONEDIR/$output_rel_path"
